standardised php response object, incorporated client-side validation, added basic authorization

+++ admin/users/admin_users.js
@@ -1,3 +1,36 @@
+import { checkAuth } from '../../utils/auth.js'
+
+function validatePassword(errObj, password) {
+  if (username && !password.match(/^.{8,}$/)) {
+    errObj.success = false
+    errObj.message += 'A password must have at least 8 characters.\n'
+  }
+
+  return errObj
+}
+
+function validateUsername(errObj, username) {
+  if (!username) {
+    errObj.success = false
+    errObj.message += 'Please enter a username.\n'
+  }
+
+  if (username && !username.match(/^[a-zA-Z0-9]+$/)) {
+    errObj.success = false
+    errObj.message += 'A username may only include letters and numbers.\n'
+  }
+
+  return errObj
+}
+
+function validateDelAdmin(errObj, username) {
+  if (username === 'admin') {
+    errObj.success = false
+    errObj.message += 'You are not allowed to delete this user.\n'
+  }
+  return errObj
+}
+
 async function getData() {
   const msgArea = document.getElementById('response-message')
 
@@ -19,10 +52,21 @@ async function getData() {
   return json.data
 }
 
-async function handleDelete(userId) {
+async function handleDelete(userData) {
   const msgArea = document.getElementById('response-message')
 
-  const response = await fetch(`./admin_users.php?userId=${userId}`, {
+  // prevent submission if name is "admin"
+  // will give multiple messages if multiple problems are detected
+  let errObj = { success: true, message: '' }
+
+  errObj = validateDelAdmin(errObj, userData.username)
+
+  if (!errObj.success) {
+    msgArea.textContent = errObj.message
+    return
+  }
+
+  const response = await fetch(`./admin_users.php?userId=${userData.userId}`, {
     method: 'DELETE',
   })
 
@@ -87,7 +131,7 @@ function renderUsers(usersData) {
     // set up del button event
     document
       .getElementById(`delBtn-${i}`)
-      .addEventListener('click', () => handleDelete(usersData[i].userId))
+      .addEventListener('click', () => handleDelete(usersData[i]))
 
     document
       .getElementById(`passBtn-${i}`)
@@ -127,6 +171,17 @@ async function handleSaveNewPassword(usersData, i) {
 
   const newPassword = document.getElementById(`input-pass-${i}`).value
 
+  // prevent submission if password isn't valid
+  // will give multiple messages if multiple problems are detected
+  let errObj = { success: true, message: '' }
+
+  errObj = validatePassword(errObj, newPassword)
+
+  if (!errObj.success) {
+    msgArea.textContent = errObj.message
+    return
+  }
+
   const updatedUsersData = { ...usersData[i], password: newPassword }
 
   const response = await fetch('./admin_users.php', {
@@ -160,10 +215,22 @@ async function handleSaveNewPassword(usersData, i) {
 async function handleSaveEditUser(usersData, i) {
   const msgArea = document.getElementById('response-message')
 
+  // gather the data from the input
   const editedUserData = {}
 
   for (const key of Object.keys(usersData[i])) {
     editedUserData[key] = document.getElementById(`input-${key}-${i}`).value
+  }
+
+  // prevent submission if name isn't valid
+  // will give multiple messages if multiple problems are detected
+  let errObj = { success: true, message: '' }
+
+  errObj = validateUsername(errObj, editedUserData.username)
+
+  if (!errObj.success) {
+    msgArea.textContent = errObj.message
+    return
   }
 
   const response = await fetch('./admin_users.php', {
@@ -209,6 +276,7 @@ function handleEditUser(usersData, i) {
         ${key === 'userId' && 'disabled'}
         ${key === 'dateJoined' && 'disabled'}
         ${key === 'isAdmin' && 'disabled'}
+        ${key === 'username' && value === 'admin' && 'disabled'}
       >`
   }
 
@@ -234,6 +302,18 @@ async function handleAddNewUser() {
   const password = document.getElementById('password').value
   const isAdminBool = document.getElementById('admin').checked
   const isAdmin = isAdminBool ? '1' : '0'
+
+  // prevent submission if name or password isn't valid
+  // will give multiple messages if multiple problems are detected
+  let errObj = { success: true, message: '' }
+
+  errObj = validateUsername(errObj, username)
+  errObj = validatePassword(errObj, password)
+
+  if (!errObj.success) {
+    msgArea.textContent = errObj.message
+    return
+  }
 
   const response = await fetch('./admin_users.php', {
     method: 'POST',
@@ -262,6 +342,9 @@ async function handleAddNewUser() {
 }
 
 async function init() {
+  // prevent unauthorized users from entering admin area
+  checkAuth(true)
+
   // set up add user
   document.getElementById('addUserBtn').addEventListener('click', handleAddNewUser)
